ref: move get_middle up to the base class so it can be shared
